Flask Deploy/requirements.txt
Flask
flask-cors
numpy
tensorflow
Pillow
h5py
